requirements.txt
tqdm~=4.66.2
numpy~=1.26.1
PyYAML~=6.0.1
utils~=1.0.2
pillow~=10.2.0